getfloat age

+++ Fantasia/Assets/Scripts/TutorialAnsia/NextButtonScript.cs
@@ -21,7 +21,7 @@ public class NextButtonScript : MonoBehaviour
 
     private int clust;
     private string username;
-    private int age;
+    private float age;
     
     private void Awake()
     {
@@ -29,7 +29,7 @@ public class NextButtonScript : MonoBehaviour
         Cursor.lockState = CursorLockMode.None;
         //clust = 2;
         clust = PlayerPrefs.GetInt("ClusterValue");
-        age = PlayerPrefs.GetInt("Age");
+        age = PlayerPrefs.GetFloat("Age");
         //age = 30;
         username = PlayerPrefs.GetString("Username");
         
@@ -83,7 +83,7 @@ public class NextButtonScript : MonoBehaviour
             "In the following game the dragon will move up and down according to the rhythm of the breath, try to follow the writes in the cloud with the arrows of your board");
         TutorialCharacterMovement.active = true;
 
-        if (age > 45)
+        if (age > 45f)
         {
             arrowup.gameObject.SetActive(true);
             arrowdown.gameObject.SetActive(true);
@@ -153,7 +153,7 @@ public class NextButtonScript : MonoBehaviour
             {
                 if(currentLineIndex == 1)
                 {
-                    if(age>45)
+                    if(age>45f)
                     {
                         indlives.gameObject.SetActive(true);
                     }
@@ -163,7 +163,7 @@ public class NextButtonScript : MonoBehaviour
                 {
                     heart3.gameObject.SetActive(false);
                     
-                    if(age>45)
+                    if(age>45f)
                     {
                       indlives.gameObject.SetActive(false);
                       indlostlife.gameObject.SetActive(true);
